requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
openai==1.14.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==4.0.7
Flask-Admin==1.6.1
Flask-Cors==4.0.0
gunicorn==21.2.0
Flask-AppBuilder==4.4.1
